Add hardware platform lookup
