+++ onnxruntime/core/mlas/lib/mlasi.h
@@ -855,6 +855,14 @@ extern const MLAS_CONV_SYM_DISPATCH MlasConvSymS8DispatchNeon;
 extern const MLAS_CONV_SYM_DISPATCH MlasConvSymU8DispatchDot;
 extern const MLAS_CONV_SYM_DISPATCH MlasConvSymS8DispatchDot;
 
+struct MLAS_Q8Q4GEMM_DISPATCH;
+
+extern const MLAS_Q8Q4GEMM_DISPATCH MlasQ8Q4GemmDispatchAvx512vnni;
+
+struct MLAS_FPQ4GEMM_DISPATCH;
+
+extern const MLAS_FPQ4GEMM_DISPATCH MlasFpQ4GemmDispatchAvx512;
+
 //
 // Quantized depthwise convolution kernels.
 //
@@ -980,6 +988,8 @@ struct MLAS_PLATFORM {
     static constexpr int32_t MaximumThreadCount = MLAS_MAXIMUM_THREAD_COUNT;
 #endif
 
+    const MLAS_FPQ4GEMM_DISPATCH* FpQ4GemmDispatch{nullptr};
+    const MLAS_Q8Q4GEMM_DISPATCH* Q8Q4GemmDispatch{nullptr};
 };
 
 inline
