+++ onnxruntime/python/tools/quantization/matmul_weight4_quantizer.py
@@ -4,7 +4,7 @@
 # license information.
 # --------------------------------------------------------------------------
 
-from typing import Tuple
+from typing import List, Tuple
 
 import numpy as np
 import onnx
@@ -34,7 +34,7 @@ class MatMulWeight4Quantizer:
         self.quant_type = quant_type
 
     @staticmethod
-    def __get_initializer(name, graph_path: list[GraphProto]) -> Tuple[TensorProto, GraphProto]:
+    def __get_initializer(name, graph_path: List[GraphProto]) -> Tuple[TensorProto, GraphProto]:
         for gid in range(len(graph_path) - 1, -1, -1):
             graph = graph_path[gid]
             for tensor in graph.initializer:
@@ -42,7 +42,7 @@ class MatMulWeight4Quantizer:
                     return tensor, graph
         return None, None
 
-    def _q4_matmul_node_weight(self, node: NodeProto, graph_stack: list[GraphProto]) -> NodeProto:
+    def _q4_matmul_node_weight(self, node: NodeProto, graph_stack: List[GraphProto]) -> NodeProto:
         """If the node is MatMul with fp32 const weight, quantize the weight with int4, and return the new node"""
 
         if node.op_type != "MatMul":
@@ -85,7 +85,7 @@ class MatMulWeight4Quantizer:
         )
         return matmul_q4_node
 
-    def _process_subgraph(self, graph_stack: list[GraphProto]):
+    def _process_subgraph(self, graph_stack: List[GraphProto]):
         new_nodes = []
         graph = graph_stack[-1]
