+++ onnxruntime/test/mlas/bench/bench_q4gemm.cpp
@@ -107,7 +107,6 @@ void Q8Q4GEMM(benchmark::State& state, MLAS_BLK_QUANT_TYPE qtype) {
   }
 }
 
-
 static void GemmSizeProducts(benchmark::internal::Benchmark* b) {
   b->ArgNames(q4gemm_bench_arg_names);
   ArgsProduct(b, {{1, 1024, 2048}, {4096}, {4096}, {8}});
